+++ zcash_client_sqlite/src/testing/pool.rs
@@ -37,6 +37,13 @@ pub(crate) fn send_single_step_proposed_transfer<T: ShieldedPoolTester>() {
     )
 }
 
+pub(crate) fn scan_full_block_detects_outputs<T: ShieldedPoolTester>() {
+    zcash_client_backend::data_api::testing::pool::scan_full_block_detects_outputs::<T>(
+        TestDbFactory::default(),
+        BlockCache::new(),
+    )
+}
+
 pub(crate) fn spend_max_spendable_single_step_proposed_transfer<T: ShieldedPoolTester>() {
     zcash_client_backend::data_api::testing::pool::spend_max_spendable_single_step_proposed_transfer::<
         T,